Update Object Storage API Reference to point to wadl in api-site
The point of this change is to ensure that
these two docs are in sync and are complete with regards
to the descriptions of each method and their request
and response parameters:
http://api.openstack.org/api-ref-objectstorage.html
(API Reference page for Object Storage)
http://docs.openstack.org/api/openstack-object-storage/1.0/content/
(API Reference (or spec) for Object Storage)
In addition, the current documents don’t describe ACLs, FormPOST, TempURLs,
StaticWeb, Bulk Upload, Bulk Delete, the OPTIONS operation
(though this is implicit in CORS support) and maybe a couple of other things.
Also, validate all code examples.

Co-Author: Donagh McCabe

Partial-Bug: #1187119

Change-Id: I94054b046a94260ba8825bdb42439adfcaf9fdce
author: diane fleming